backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o backend/requirements.txt
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
django==5.1.7
    # via
    #   discocatdjango (pyproject.toml)
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-admin==2.0.2
    # via discocatdjango (pyproject.toml)
django-cors-headers==4.7.0
    # via discocatdjango (pyproject.toml)
django-excel-response2==3.0.6
    # via django-admin
django-filter==25.1
    # via discocatdjango (pyproject.toml)
django-six==1.0.5
    # via
    #   django-admin
    #   django-excel-response2
djangorestframework==3.15.2
    # via discocatdjango (pyproject.toml)
excel-base==1.0.4
    # via django-excel-response2
gunicorn==23.0.0
    # via discocatdjango (pyproject.toml)
isoweek==1.3.3
    # via timeconvert
packaging==24.2
    # via gunicorn
pillow==11.1.0
    # via discocatdjango (pyproject.toml)
python-dateutil==2.9.0.post0
    # via timeconvert
python-dotenv==1.0.1
    # via discocatdjango (pyproject.toml)
screen==1.0.1
    # via excel-base
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
timeconvert==3.0.13
    # via excel-base
tzlocal==5.3.1
    # via timeconvert
xlwt==1.3.0
    # via excel-base
